refactor(backend): removed core/, added models/
