+++ new/src/backend/app/api/v1/exceptions.py
