+++ new/src/backend/app/api/v1/middleware/user_loader.py
@@ -3,13 +3,13 @@
 from typing import Awaitable, Callable
 
 
-from core.dependency_injection.container import Container
-from core.logging.base import LoggerBase
+from api.v1.container import Container
 from db.repositories.user_repository import UserRepository
 from starlette.middleware.base import BaseHTTPMiddleware
 from starlette.requests import Request
 from starlette.responses import Response
 
+from services.logging.base import LoggerBase
 from services.oidc import OIDCService
 
 RequestResponseEndpoint = Callable[[Request], Awaitable[Response]]
@@ -38,6 +38,10 @@ class UserLoaderMiddleware(BaseHTTPMiddleware):
                 self.logger.info(f"Request contains valid token, loading user {openid}.")
                 user = self.user_repository.get(openid)
                 request.state.user = user
+
+        # openid = "f8e916e578ec64cbb26786c2769b4a95bd87aef0@lifescience-ri.eu"
+        # user = self.user_repository.get(openid)
+        # request.state.user = user
 
         response = await call_next(request)
         return response
