+++ new/src/backend/app/main.py
@@ -5,8 +5,8 @@ import shutil
 
 from dotenv import load_dotenv
 from fastapi import FastAPI, HTTPException
-from typing import Tuple
 
+from api.v1.container import Container
 from api.v1.routes.charges import charges_router
 from api.v1.routes.auth import auth_router
 from api.v1.routes.files import files_router
@@ -14,16 +14,11 @@ from api.v1.middleware.logging import LoggingMiddleware
 from api.v1.middleware.exceptions import http_exception_handler
 from api.v1.middleware.user_loader import UserLoaderMiddleware
 
-from core.dependency_injection.container import Container
-
 
 PREFIX = "/v1"
-WEB_PREFIX = f"{PREFIX}"
 
 
-def create_app() -> FastAPI:
-    """Creates FastAPI aps with routers and middleware."""
-
+def move_example_files() -> None:
     if not load_dotenv():
         raise EnvironmentError("Could not load environment variables.")
 
@@ -31,16 +26,25 @@ def create_app() -> FastAPI:
     if not (examples_dir := os.environ.get("ACC2_EXAMPLES_DIR")):
         raise EnvironmentError("ACC2_EXAMPLES_DIR environment variable is not set.")
 
+    if os.path.exists(examples_dir):
+        shutil.rmtree(examples_dir)
+
     try:
         shutil.copytree("examples", examples_dir, dirs_exist_ok=True)
     except Exception as e:
         raise EnvironmentError(f"Could not copy example files. {str(e)}") from e
 
+
+def create_app() -> FastAPI:
+    """Creates FastAPI aps with routers and middleware."""
+
+    move_example_files()
+
     # Create DI container
     container = Container()
 
-
     app = FastAPI(root_path="/api")
+    # app = FastAPI()
 
     container.wire()
