+++ new/src/backend/app/models/calculation.py
@@ -7,8 +7,8 @@ import uuid
 from pydantic import BaseModel, ConfigDict
 from pydantic.alias_generators import to_camel
 
-from core.models.paging import PagingFilters
-from core.models.molecule_info import MoleculeSetStats
+from models.paging import PagingFilters
+from models.molecule_info import MoleculeSetStats
 
 from integrations.chargefw2.base import Charges
