+++ new/src/backend/app/models/molecule_info.py
