+++ new/src/backend/app/services/io.py
@@ -10,10 +10,10 @@ from typing import Tuple
 from dotenv import load_dotenv
 from fastapi import UploadFile
 
-from core.logging.base import LoggerBase
-from core.models.calculation import ChargeCalculationConfigDto
+from models.calculation import ChargeCalculationConfigDto
 
 from integrations.io.base import IOBase
+from services.logging.base import LoggerBase
 
 load_dotenv()
